simulacrum + DataFrame.apply
