+++ arrow/src/main/scala/org/sciscala/simpledf/arrow/implicits.scala
@@ -1,0 +1,9 @@
+package org.sciscala.simpledf.arrow
+
+import org.sciscala.simpledf.DataFrame
+
+object implicits {
+
+  implicit val arrowDataFrame: DataFrame[ArrowDataFrame] = DataFrame.apply(ArrowDataFrame.dfInstance)
+
+}
