+++ arrow/src/test/scala/org/sciscala/simpledf/arrow/ArrowDataFrameSpec.scala
@@ -16,9 +16,11 @@ import org.scalactic.Equality
 import scala.collection.mutable.ListBuffer
 
 import org.sciscala.simpledf._
-import org.sciscala.simpledf.arrow._
+import org.sciscala.simpledf.DataFrame.ops._
+import org.sciscala.simpledf.arrow.ArrowDataFrame
+import org.sciscala.simpledf.arrow.ArrowDataFrame._
 import org.sciscala.simpledf.arrow.ArrowEncoder._
-import org.sciscala.simpledf.arrow.Implicits._
+import org.sciscala.simpledf.arrow.implicits._
 import scala.languageFeature.postfixOps
 
 class ArrowDataFrameSpec extends AnyFlatSpec with Matchers {
@@ -46,6 +48,7 @@ class ArrowDataFrameSpec extends AnyFlatSpec with Matchers {
     }
 
   }
+
 
   implicit val schemaRootDataFrame: Equality[VectorSchemaRoot] =
     (a: VectorSchemaRoot, b: Any) => b match {
@@ -166,56 +169,56 @@ class ArrowDataFrameSpec extends AnyFlatSpec with Matchers {
   )
 
   "Size" should "equal 12" in {
-    DataFrame[ArrowDataFrame].size(df) shouldEqual 12
+    df.size shouldEqual 12
   }
 
   "Shape" should "equal (6,2)" in {
-    DataFrame[ArrowDataFrame].shape(df) shouldBe (6, 2)
+    df.shape shouldBe (6,2)
   }
 
   "Head" should "return a Dataframe with the first 5 rows" in {
-    val d = DataFrame[ArrowDataFrame].head(df)
+    val d = df.head()
     d.data shouldEqual data.slice(0, 5)
     d.index shouldBe index.take(5)
     d.columns shouldBe cols.take(5)
   }
 
   "head(2)" should "return a Dataframe with the first 2 rows" in {
-    val d = DataFrame[ArrowDataFrame].head(df, 2)
+    val d = df.head(2)
     d.data shouldEqual data.slice(0, 2)
     d.index shouldBe index.take(2)
     d.columns shouldBe cols.take(2)
   }
 
   "Head of dfNoIndex" should "return a Dataframe with the first 5 rows and an empty index" in {
-    val d = DataFrame[ArrowDataFrame].head(dfNoIndex)
+    val d = dfNoIndex.head()
     d.data shouldEqual data.slice(0, 5)
     d.index shouldBe ArraySeq()
     d.columns shouldBe cols.take(5)
   }
 
   "iat(2,1)" should "equal 8" in {
-    DataFrame[ArrowDataFrame].iat(df, 2, 1) shouldBe Some(8)
+    df.iat(2,1) shouldBe Some(8)
   }
 
   "iat(-2,1)" should "equal 14" in {
-    DataFrame[ArrowDataFrame].iat(df, -2, 1) shouldBe Some(14)
+    df.iat(-2,1) shouldBe Some(14)
   }
 
   "iat(2,-1)" should "equal 7" in {
-    DataFrame[ArrowDataFrame].iat(df, 2, -1) shouldBe Some(8)
+    df.iat(2,-1) shouldBe Some(8)
   }
 
   "iat(-2,-1)" should "equal 13" in {
-    DataFrame[ArrowDataFrame].iat(df, -2, -2) shouldBe Some(13)
+    df.iat(-2,-2) shouldBe Some(13)
   }
 
   "at('viper', 'stamina')" should "equal 2" in {
-    DataFrame[ArrowDataFrame].at(df, "viper", "stamina") shouldBe Some(2)
+    df.at("viper", "stamina") shouldBe Some(2)
   }
 
   "at('viper', 'venom')" should "be None" in {
-    DataFrame[ArrowDataFrame].at(df, "viper", "venom") shouldBe None
+    df.at("viper", "venom") shouldBe None
   }
 
   "get(name)" should "return the dataframe with the `name` label" in {
@@ -235,7 +238,7 @@ class ArrowDataFrameSpec extends AnyFlatSpec with Matchers {
     data.setRowCount(6)
 
     val expected = ArrowDataFrame(data, index)
-    val arrDF = DataFrame[ArrowDataFrame].get[String](df, "speed", None)
+    val arrDF = df.get[String]("speed", None)
 
     arrDF.data.getFieldVectors.size() shouldBe 1
     val actual = arrDF.data.getVector(0)
@@ -244,18 +247,18 @@ class ArrowDataFrameSpec extends AnyFlatSpec with Matchers {
   }
 
   "get(unknown)" should "return the default value" in {
-    DataFrame[ArrowDataFrame].get[String](df, "poison", None) shouldBe nullArrowDF
+    df.get[String]("poison", None) shouldBe emptyInstance
   }
 
   "loc('viper')" should "return df's first row" in {
     data.clear()
     data.allocateNew()
     val expected = serpents2DF(Seq(Serpent("viper", 1, 2)), data)
-    DataFrame[ArrowDataFrame].loc(df, "viper") shouldBe expected
+    df.loc("viper") shouldBe expected
   }
 
   "loc('venom')" should "return 'null' dataframe" in {
-    DataFrame[ArrowDataFrame].loc(df, "venom") shouldEqual nullArrowDF
+    df.loc("venom") shouldEqual emptyInstance
   }
 
   "loc('viper', 'python')" should "return first and fourth rows" in {
@@ -267,7 +270,7 @@ class ArrowDataFrameSpec extends AnyFlatSpec with Matchers {
     )
     initData(data)
     val wholeSet = ArrowDataFrame(data, index)
-    val res = DataFrame[ArrowDataFrame].loc(wholeSet, Seq("viper", "python"))
+    val res = wholeSet.loc(Seq("viper", "python"))
     println(expected.data.toString())
     res.data shouldEqual expected.data
     res.index shouldBe expected.index
@@ -277,7 +280,7 @@ class ArrowDataFrameSpec extends AnyFlatSpec with Matchers {
     data.clear()
     data.allocateNew()
     val df = serpents2DF(Seq(Serpent("viper", 1, 2)), data)
-    val res = DataFrame[ArrowDataFrame].loc(df, Seq("viper", "venom"))
+    val res = wholeSet.loc(Seq("viper", "venom"))
     res.data shouldEqual df.data
     res.index shouldBe df.index
   }
@@ -291,8 +294,7 @@ class ArrowDataFrameSpec extends AnyFlatSpec with Matchers {
     )
     initData(data)
     val wholeSet = ArrowDataFrame(data, index)
-    val res = DataFrame[ArrowDataFrame]
-      .loc(wholeSet, Seq(true, false, false, true, false, false))
+    val res = wholeSet.loc(Seq(true, false, false, true, false, false))
     res.data shouldEqual expected.data
     res.index shouldBe expected.index
   }
@@ -300,8 +302,7 @@ class ArrowDataFrameSpec extends AnyFlatSpec with Matchers {
   "loc(false, false, false, false, false, false)" should "return a dataframe with no rows and only colNames" in {
     initData(data)
     val wholeSet = ArrowDataFrame(data, index)
-    val res = DataFrame[ArrowDataFrame]
-      .loc(wholeSet, Seq(false, false, false, false, false, false))
+    val res = wholeSet.loc(Seq(false, false, false, false, false, false))
     res.data.getRowCount shouldEqual 0
     res.index.size shouldBe 0
   }
@@ -309,8 +310,7 @@ class ArrowDataFrameSpec extends AnyFlatSpec with Matchers {
   "loc(true, true, true, true, true, true)" should "be the original dataframe" in {
     initData(data)
     val wholeSet = ArrowDataFrame(data, index)
-    val res = DataFrame[ArrowDataFrame]
-      .loc(wholeSet, Seq(true, true, true, true, true, true))
+    val res = wholeSet.loc(Seq(true, true, true, true, true, true))
     res.data shouldEqual df.data
     res.index shouldBe df.index
   }
@@ -328,19 +328,19 @@ class ArrowDataFrameSpec extends AnyFlatSpec with Matchers {
   }
 
   "to" should "encode DataFrame" in {
-    DataFrame[ArrowDataFrame].to(wholeSet) shouldBe serpents
+    wholeSet.to[Serpent] shouldBe serpents
   }
 
   "to" should "encode DataFrame with no index" in {
-    DataFrame[ArrowDataFrame].to(dfNoIndex) shouldBe serpentsNoIndex
+    dfNoIndex.to[Serpent] shouldBe serpentsNoIndex
   }
 
   "empty" should "return true if dataframe is empty" in {
-    DataFrame[ArrowDataFrame].empty(nullArrowDF) shouldBe true
+    emptyInstance.empty shouldBe true
   }
 
   "empty" should "return false if dataframe is not empty" in {
-    DataFrame[ArrowDataFrame].empty(df) shouldBe false
+    df.empty shouldBe false
   }
 
   "insert" should "adds a column at index `loc`" in {
@@ -360,20 +360,20 @@ class ArrowDataFrameSpec extends AnyFlatSpec with Matchers {
     poisonVector.setSafe(4, 13)
     poisonVector.setSafe(5, 16)
 
-    val newDF = DataFrame[ArrowDataFrame]
-      .insert[UInt4Vector](df, 1, "poison", poisonVector, false)
-      .getOrElse(nullArrowDF)
+    val newDF = df
+      .insert[UInt4Vector](1, "poison", poisonVector, false)
+      .getOrElse(emptyInstance)
     newDF.data.getVector("poison") shouldBe poisonVector
   }
 
   "items" should "return data in Array[(columnName, Column)] format" in {
     val speedVectorAsSeq = ArrowUtils.vectorAsSeq(df.shape._1,data.getVector("speed"))
     val staminaVectorAsSeq = ArrowUtils.vectorAsSeq(df.shape._1,data.getVector("stamina"))
-    DataFrame[ArrowDataFrame].items(df) shouldBe Array("speed" -> speedVectorAsSeq, "stamina" -> staminaVectorAsSeq)
+    df.items shouldBe Array("speed" -> speedVectorAsSeq, "stamina" -> staminaVectorAsSeq)
   }
 
   "items" should "return empty array for emptyDF" in {
-    DataFrame[ArrowDataFrame].items(nullArrowDF) shouldBe Array() 
+    emptyInstance.items shouldBe Array() 
   }
 
   "iterrows" should "return sequence of tuples (index, Row) format" in {
@@ -384,7 +384,8 @@ class ArrowDataFrameSpec extends AnyFlatSpec with Matchers {
       Field("speed", IntType, false),
       Field("stamina", IntType, false)
     ))
-    DataFrame[ArrowDataFrame].iterrows(df)(arrowRowEncoder(dfSchema)) shouldBe Seq(
+    implicit val rowEncoder: Encoder[ArrowDataFrame, Row] = arrowRowEncoder(dfSchema)
+    df.iterrows shouldBe Seq(
       ("viper",Row(Seq(1,2), dfSchema)),
       ("sidewinder",Row(Seq(4,5), dfSchema)),
       ("cobra",Row(Seq(7,8), dfSchema)),
@@ -398,11 +399,13 @@ class ArrowDataFrameSpec extends AnyFlatSpec with Matchers {
 
   "iterrows" should "return empty sequence of tuples for emptyDF" in {
     import org.sciscala.simpledf.types._
+    import org.sciscala.simpledf.row.Row
     val dfSchema = Schema(Seq(
       Field("name", StringType, false),
       Field("speed", IntType, false),
       Field("stamina", IntType, false)
     ))
-    DataFrame[ArrowDataFrame].iterrows(nullArrowDF)(arrowRowEncoder(dfSchema)) shouldBe Seq()
+    implicit val rowEncoder: Encoder[ArrowDataFrame, Row] = arrowRowEncoder(dfSchema)
+    emptyInstance.iterrows shouldBe Seq()
   }
 }
