+++ build.sbt
@@ -11,7 +11,7 @@ lazy val core = (project in file("core"))
     name := "simpledf-core",
     scalacOptions += "-Ymacro-annotations",
     resolvers += "Sonatype OSS Snapshots" at "https://oss.sonatype.org/content/repositories/snapshots",
-    libraryDependencies += simulacrum, //TODO: Move to simulacrum-scalafix
+    libraryDependencies += "org.typelevel" %% "simulacrum" % "1.0.0", // TODO: Move to simulacrum-scalafix
     libraryDependencies += "org.typelevel" %% "spire" % "0.17.0-M1",
     //libraryDependencies += "io.estatico" %% "newtype" % "0.4.4",
     libraryDependencies += scalaTest % Test
@@ -22,7 +22,7 @@ lazy val arrow = (project in file("arrow"))
   .settings(
     name := "simpledf-arrow",
     scalacOptions += "-Ymacro-annotations",
-    libraryDependencies += simulacrum, //TODO: Move to simulacrum-scalafix
+    libraryDependencies += "org.typelevel" %% "simulacrum" % "1.0.0", // TODO: Move to simulacrum-scalafix
     libraryDependencies += "org.typelevel" %% "spire" % "0.17.0-M1",
     libraryDependencies += "org.apache.arrow" % "arrow-vector" % "0.17.1",
     libraryDependencies += scalaTest % Test
